# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Meter API",
        "version": "1.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur local"
        },
        {
            "url": "http://localhost:8000",
            "description": "Serveur local"
        },
        {
            "url": "http://localhost:8000",
            "description": "Serveur local"
        },
        {
            "url": "http://localhost:8000",
            "description": "Serveur local"
        },
        {
            "url": "http://localhost:8000",
            "description": "Serveur local"
        },
        {
            "url": "http://localhost:8000",
            "description": "Serveur local"
        },
        {
            "url": "http://localhost:8000",
            "description": "Serveur local"
        }
    ],
    "paths": {
        "/api/compteurs/user/{id}": {
            "get": {
                "tags": [
                    "Compteurs"
                ],
                "summary": "Obtenir les compteurs d'un utilisateur",
                "operationId": "2b47c7f818771345464f73d7c3848129",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des compteurs lié a un utilisateur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "Compteur"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Utilisateur non trouvé"
                    }
                }
            }
        },
        "/api/compteurs/new": {
            "post": {
                "tags": [
                    "Compteurs"
                ],
                "summary": "Ajoute un nouveau compteur",
                "operationId": "f325b31125d0f9cab2644d1488e81510",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nummero_compteur",
                                    "localisation",
                                    "type_local",
                                    "frequence_moy_rechargement",
                                    "user_id"
                                ],
                                "properties": {
                                    "nummero_compteur": {
                                        "type": "string",
                                        "example": "123456"
                                    },
                                    "localisation": {
                                        "type": "string",
                                        "example": "123 Rue Example"
                                    },
                                    "type_local": {
                                        "type": "string",
                                        "example": "Appartement"
                                    },
                                    "frequence_moy_rechargement": {
                                        "type": "integer",
                                        "example": 30
                                    },
                                    "montant_moy_rechargement": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 50.5
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Compteur créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "succès"
                                        },
                                        "data": {
                                            "$ref": "schemas/Compteur"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erreur de validation"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/compteurs/update/{id}": {
            "put": {
                "tags": [
                    "Compteurs"
                ],
                "summary": "Met à jour un compteur existant",
                "operationId": "0a7cf705f9a357239d453b3656e756bd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du compteur à mettre à jour",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "numero_compteur": {
                                        "type": "string",
                                        "example": "123456"
                                    },
                                    "localisation": {
                                        "type": "string",
                                        "example": "123 Rue Example"
                                    },
                                    "type_local": {
                                        "type": "string",
                                        "example": "Appartement"
                                    },
                                    "frequence_moy_rechargement": {
                                        "type": "integer",
                                        "example": 30
                                    },
                                    "montant_moy_rechargement": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 50.5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Compteur mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Mise à jour réussie"
                                        },
                                        "data": {
                                            "$ref": "schemas/Compteur"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erreur de validation"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Compteur non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Compteur non trouvé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}